chore: create basic toolbar component
